--- cpg-core/src/test/resources/lambda.cpp
@@ -1,5 +1,0 @@
-int main() {
-SSL_CTX_set_verify(ctx, SSL_VERIFY_PEER, [](int preverify_ok, X509_STORE_CTX *x509_ctx) {
-    return 1;
-  });
-}
